requirements.txt
anyio==3.3.4
asgiref==3.4.1
black==21.12b0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
Flask==2.0.2
Flask-Login==0.5.0
Flask-WTF==1.0.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
pdfkit==1.0.0
Pillow==8.4.0
platformdirs==2.4.1
pydantic==1.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.3
typing-extensions==3.10.0.2
uvicorn==0.15.0
Werkzeug==2.0.2
WTForms==3.0.1
